target label dtype changed

+++ dataset.py
@@ -10,7 +10,7 @@ class EvcDataset(Dataset):
         self.hs = torch.tensor(hs).float()
         self.ts = torch.tensor(ts).int()  # keep int dtype -> goes to embedding layer
         self.ss = torch.tensor(ss).float()
-        self.ys = torch.tensor(ys).float()
+        self.ys = torch.tensor(ys).to(torch.int64)
 
     def __len__(self):
         return len(self.rs)
